pyproject.toml
[tool.poetry]
name = "prime_number_finder_backend"
version = "0.1.0"
description = "Previous Prime Number Calculator"
authors = ["Catrius Pham <catriuspham@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.1"
factory-boy = "^3.0.1"
factory_boy = "^3.0.1"
django-environ = "^0.4.5"
django-cors-headers = "^3.5.0"
psycopg2-binary = "^2.8.6"
djangorestframework = "^3.11.1"
coverage = "^5.3"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
sure = "^1.4.11"
ipdb = "^0.13.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
